% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antiword.R
\name{antiword}
\alias{antiword}
\title{Antiword}
\usage{
antiword(file = NULL, format = FALSE)
}
\arguments{
\item{file}{path or url to your word file}

\item{format}{format the output text (-f parameter)}
}
\description{
Wraps the antiword utility. Takes a path to an word file and returns
text from the document.
}
\examples{
text <- antiword("https://jeroen.github.io/files/UDHR-english.doc")
cat(text)
}
